Renamed general string utils file to be more specific
